requirements.txt
langchain==0.1.14
langchain-community==0.0.30
huggingface-hub==0.23.1
transformers==4.40.1
sentence-transformers==2.7.0
qdrant-client==1.7.0
python-dotenv==1.0.1
pandas==2.2.2
tqdm==4.66.2
scikit-learn>=1.4.0
numpy>=1.24
